requirements.txt
slick-bitcoinrpc==0.1.4
python-dotenv==0.21.1
requests==2.31.0
